Correct existing data parsing and JSON export functionality
This commit addresses several bugs in the way the script was parsing incoming data and exporting it as JSON. Notably, the '.replace' method was corrected to cater for semicolons in 'db_name' and 'table' values. Furthermore, JSON filtering was adjusted to return all objects, rather than those with both 'db_name' and 'table' specifically. Lastly, JSON output is now wrapped in an object for better structure and readability.